+++ GrappleECS/src/GrappleECS/Query/EntityView.cpp
@@ -13,7 +13,7 @@ namespace Grapple
 		return EntityViewIterator(storage, storage.GetEntitiesCount());
 	}
 
-	std::optional<Entity> EntityView::GetEntity(size_t index)
+	std::optional<Entity> EntityView::GetEntity(uint32_t index)
 	{
 		const auto& indices = m_Registry.GetArchetypeRecord(m_Archetype).Storage.GetEntityIndices();
 		Grapple_CORE_ASSERT(index < indices.size());
