+++ GrappleECS/src/GrappleECS/Query/EntityView.h
@@ -21,7 +21,7 @@ namespace Grapple
 		EntityViewIterator begin();
 		EntityViewIterator end();
 
-		std::optional<Entity> GetEntity(size_t index);
+		std::optional<Entity> GetEntity(uint32_t index);
 
 		constexpr ArchetypeId GetArchetype() const { return m_Archetype; }
