+++ GrappleScriptingCore/src/GrappleScriptingCore/Bindings/ECS/Component.h
@@ -4,6 +4,6 @@
 
 namespace Grapple::Internal
 {
-	using ComponentId = size_t;
-	constexpr ComponentId INVALID_COMPONENT_ID = SIZE_MAX;
+	using ComponentId = uint32_t;
+	constexpr ComponentId INVALID_COMPONENT_ID = UINT32_MAX;
 }
